requirements.txt
dune-client>=1.9.3
python-dotenv>=0.21.0
pyyaml>=6.0
